[REF] sale_[timesheet,project]: split model per file
Each of those modules now contains a folder `models`,
which contains one file per odoo model.
