+++ addons/sale_project/models/__init__.py
@@ -1,10 +1,11 @@
-# -*- coding: utf-8 -*-
-
-from . import account_move
-from . import product
-from . import project
+from . import account_move_line
+from . import product_product
+from . import product_template
 from . import project_milestone
-from . import sale_order
+from . import project_project
+from . import project_task_recurrence
+from . import project_task
+from . import res_config_settings
 from . import sale_order_line
 from . import sale_order_template_line
-from . import res_config_settings
+from . import sale_order
