+++ addons/sale_project/models/product_template.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from odoo import api, fields, models, _
@@ -167,7 +166,7 @@ class ProductTemplate(models.Model):
 
     @api.onchange('type')
     def _onchange_type(self):
-        res = super(ProductTemplate, self)._onchange_type()
+        res = super()._onchange_type()
         if self.type != 'service':
             self.service_tracking = 'no'
         return res
@@ -178,39 +177,4 @@ class ProductTemplate(models.Model):
                 'service_tracking': 'no',
                 'project_id': False
             })
-        return super(ProductTemplate, self).write(vals)
-
-
-class ProductProduct(models.Model):
-    _inherit = 'product.product'
-
-    @api.onchange('service_tracking')
-    def _onchange_service_tracking(self):
-        if self.service_tracking == 'no':
-            self.project_id = False
-            self.project_template_id = False
-        elif self.service_tracking == 'task_global_project':
-            self.project_template_id = False
-        elif self.service_tracking in ['task_in_project', 'project_only']:
-            self.project_id = False
-
-    def _inverse_service_policy(self):
-        for product in self:
-            if product.service_policy:
-
-                product.invoice_policy, product.service_type = self.product_tmpl_id._get_service_to_general(product.service_policy)
-
-    @api.onchange('type')
-    def _onchange_type(self):
-        res = super(ProductProduct, self)._onchange_type()
-        if self.type != 'service':
-            self.service_tracking = 'no'
-        return res
-
-    def write(self, vals):
-        if 'type' in vals and vals['type'] != 'service':
-            vals.update({
-                'service_tracking': 'no',
-                'project_id': False
-            })
-        return super(ProductProduct, self).write(vals)
+        return super().write(vals)
