+++ addons/sale_timesheet/models/__init__.py
@@ -1,12 +1,17 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import account
+from . import account_move_line
 from . import account_move
-from . import product
-from . import project
-from . import project_update
-from . import sale_order
-from . import res_config_settings
-from . import project_sale_line_employee_map
 from . import hr_employee
+from . import hr_timesheet
+from . import product_product
+from . import product_template
+from . import project_project
+from . import project_sale_line_employee_map
+from . import project_task_recurrency
+from . import project_task
+from . import project_update
+from . import res_config_settings
+from . import sale_order_line
+from . import sale_order
